+++ src/Attribute/AttributeConfigFor.php
@@ -1,0 +1,29 @@
+<?php
+declare(strict_types=1);
+/**
+ *
+ * This file is part of Aura for PHP.
+ *
+ * @license http://opensource.org/licenses/MIT MIT
+ *
+ */
+
+namespace Aura\Di\Attribute;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS)]
+class AttributeConfigFor
+{
+    private string $className;
+
+    public function __construct(string $className)
+    {
+        $this->className = $className;
+    }
+
+    public function getClassName(): string
+    {
+        return $this->className;
+    }
+}