+++ src/ClassScanner/MapGeneratorInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Aura\Di\ClassScanner;
+
+interface MapGeneratorInterface
+{
+    public function generate(): ClassMap;
+
+    public function update(ClassMap $classMap, array $updatedFiles): ClassMap;
+}