+++ tests/ContainerBuilderTest.php
@@ -1,6 +1,8 @@
 <?php
 namespace Aura\Di;
 
+use Aura\Di\ClassScanner\ClassScannerConfig;
+use Aura\Di\ClassScanner\ComposerMapGenerator;
 use Aura\Di\Fake\FakeConstructAttributeClass;
 use Aura\Di\Fake\FakeInjectAnnotatedWithClass;
 use Aura\Di\Fake\FakeInterfaceClass1;
@@ -168,9 +170,8 @@ class ContainerBuilderTest extends TestCase
             new \Aura\Di\Fake\FakeLibraryConfig,
             new \Aura\Di\Fake\FakeProjectConfig,
             new \Aura\Di\Fake\FakeCompilationTestConfig(),
-            new ClassScanner(
-                [__DIR__ . '/Fake'],
-                ['Aura\Di\Fake'],
+            new ClassScannerConfig(
+                new ComposerMapGenerator([__DIR__ . '/Fake']),
             )
         ];